Build only for Scala-2.11 and bump version to 0.4.0
- Cleaned the build script
- Merged the tests that were running only on Scala-2.11 in the
  main test target
- Updated the README with explanation of the different versions
  and improved the formatting
