+++ src/test/scala/com/romix/akka/serialization/kryo/AkkaKryoPersistenceTests.scala
